Add block Vuex module
