+++ src/store/index.js
@@ -5,6 +5,7 @@ import moment from 'moment'
 /* MODULES */
 import auth from './modules/auth'
 import assessments from './modules/assessments'
+import blocks from './modules/blocks'
 import schedule from './modules/schedule'
 import addAssignmentModal from './modules/addAssignmentModal'
 import addExamModal from './modules/addExamModal'
@@ -26,6 +27,7 @@ export default new Vuex.Store({
   modules: {
     auth,
     assessments,
+    blocks,
     schedule,
     addAssignmentModal,
     addExamModal,
@@ -68,7 +70,7 @@ export default new Vuex.Store({
           end: moment(p.end, 'Hmm', true)
         }))
         .concat(
-          getters.getWorkBlocksAsEvents
+          getters.getAssessmentBlocksAsEvents
             .filter(e => moment(e.start).isSame(state.now, 'day'))
             .map(e => ({
               eventType: 'work-block',
